+++ src/languages.js
